# Refactor codegeneration