--- a/src/HomotopyContinuation.jl
+++ b/src/HomotopyContinuation.jl
@@ -84,8 +84,8 @@ function __init__()
     end
 end
 
-include("precompile.jl")
-_precompile_()
+# include("precompile.jl")
+# _precompile_()
 
 
 end #module