--- a/src/HomotopyContinuation.jl
+++ b/src/HomotopyContinuation.jl
@@ -84,7 +84,7 @@ function __init__()
     end
 end
 
- include("precompile.jl")
+include("precompile.jl")
 _precompile_()