--- a/src/ModelKit.jl
+++ b/src/ModelKit.jl
@@ -62,12 +62,15 @@ const MP = MultivariatePolynomials
 
 include("model_kit/symengine.jl")
 include("model_kit/symbolic.jl")
+
+include("model_kit/taylor_vector.jl")
+
 include("model_kit/instructions.jl")
-include("model_kit/taylor.jl")
-include("model_kit/abstract_types.jl")
-include("model_kit/compiler.jl")
-include("model_kit/compiled.jl")
-include("model_kit/interpreter.jl")
-include("model_kit/interpreted.jl")
+include("model_kit/slp_interpreter.jl")
+include("model_kit/slp_compiler.jl")
+#
+include("model_kit/abstract_system_homotopy.jl")
+include("model_kit/compiled_system_homotopy.jl")
+include("model_kit/interpreted_system_homotopy.jl")
 
 end # module