… src/model_kit/abstract_system_homotopy.jl
… src/model_kit/compiler.jl
… src/model_kit/interpreter.jl
… src/model_kit/slp_compiler.jl
… src/model_kit/slp_interpreter.jl
… src/model_kit/symengine.jl
… src/model_kit/taylor.jl
… src/model_kit/taylor_vector.jl
… src/newton_corrector.jl
… src/precompile.jl
… src/systems/affine_chart_system.jl
… src/systems/composition_system.jl
… src/systems/randomized_system.jl
… src/utils.jl
… test/model_kit/instructions_test.jl
… test/model_kit/slp_test.jl
… test/model_kit/symbolic_test.jl
… test/model_kit_test.jl
… test/runtests.jl
… test/solve_test.jl