This PR rewrites our code generation logic and substantially improves the `Interpreter` (see benchmarks below). In particular the interpreter performance for the taylor series computation improved significantly which is very nice since master switched in #399 to using by default the interpreter for the taylor series computations. The motivation for #399 was the realisation that the numerical derivatives are otherwise not stable in ill-conditioned situations (as during the endgame).

I also added support for `sin` and `cos` in `ModelKit` which allows to model more systems. However, this is not fully tested for now, so I wouldn't advertise this feature currently.

## Benchmark Steiner

Setup:
```julia
f = steiner();
IF = InterpretedSystem(f);
CF = CompiledSystem(f);

u = zeros(ComplexF64, 15);
U = zeros(ComplexF64, 15, 15);

y = randn(ComplexF64, 15);
p = randn(ComplexF64, 30);
x = TaylorVector{2}(randn(ComplexF64, 2, 15));
q = TaylorVector{2}(randn(ComplexF64, 2, 40));
x3 = TaylorVector{3}(randn(ComplexF64, 3, 15));
```

This PR:
```julia
julia> @btime evaluate!($u, $IF, $y, $p);
  668.557 ns (0 allocations: 0 bytes)

julia> @btime evaluate!($u, $CF, $y, $p);
  75.724 ns (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $IF, $y, $p);
  1.839 μs (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $CF, $y, $p);
  272.838 ns (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(2)), $IF, $x, $q);
  1.173 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(2)), $CF, $x, $q);
  230.872 ns (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(3)), $IF, $x3, $q);
  1.699 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(3)), $CF, $x3, $q);
  537.402 ns (0 allocations: 0 bytes)
```

Master:
```julia
julia> @btime evaluate!($u, $IF, $y, $p);
  927.677 ns (0 allocations: 0 bytes)

julia> @btime evaluate!($u, $CF, $y, $p);
  84.780 ns (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $IF, $y, $p);
  2.657 μs (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $CF, $y, $p);
  256.203 ns (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(2)), $IF, $x, $q);
  5.832 μs (4 allocations: 288 bytes)

julia> @btime taylor!($u, $(Val(2)), $CF, $x, $q);
  269.330 ns (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(3)), $IF, $x3, $q);
  9.291 μs (3 allocations: 176 bytes)

julia> @btime taylor!($u, $(Val(3)), $CF, $x3, $q);
  319.598 ns (0 allocations: 0 bytes)
```

## Benchmark Fano

Setup:
```julia
f, _ = fano_quintic()
IF = InterpretedSystem(f)
CF = CompiledSystem(f)

u = zeros(ComplexF64, 6)
U = zeros(ComplexF64, 6, 6)

y = randn(ComplexF64, 6)
p = randn(ComplexF64, 125)
x = TaylorVector{2}(randn(ComplexF64, 2, 6))
q = TaylorVector{2}(randn(ComplexF64, 2, 125))
x3 = TaylorVector{3}(randn(ComplexF64, 3, 6))
```

This PR:
```julia
julia> @btime evaluate!($u, $IF, $y, $p);
  4.697 μs (0 allocations: 0 bytes)

julia> @btime evaluate!($u, $CF, $y, $p);
  642.288 ns (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $IF, $y, $p);
  10.220 μs (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $CF, $y, $p);
  1.422 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(2)), $IF, $x, $q);
  8.681 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(2)), $CF, $x, $q);
  2.547 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(3)), $IF, $x3, $q);
  11.662 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(3)), $CF, $x3, $q);
  4.417 μs (0 allocations: 0 bytes)
```


Master:
```julia
julia> @btime evaluate!($u, $IF, $y, $p);
  13.840 μs (0 allocations: 0 bytes)

julia> @btime evaluate!($u, $CF, $y, $p);
  697.729 ns (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $IF, $y, $p);
  27.550 μs (0 allocations: 0 bytes)

julia> @btime evaluate_and_jacobian!($u, $U, $CF, $y, $p);
  1.478 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(2)), $IF, $x, $q);
  45.203 μs (4 allocations: 288 bytes)

julia> @btime taylor!($u, $(Val(2)), $CF, $x, $q);
  2.727 μs (0 allocations: 0 bytes)

julia> @btime taylor!($u, $(Val(3)), $IF, $x3, $q);
  68.610 μs (3 allocations: 176 bytes)

julia> @btime taylor!($u, $(Val(3)), $CF, $x3, $q);
  4.311 μs (0 allocations: 0 bytes)
```